Changed monotonic_ns() with monotonic() (#163)
Summary:
## Types of changes

- [ ] Docs change / refactoring / dependency upgrade
- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Motivation and Context / Related issue

The library is supposed to be compatible with Python 3.6 but one method in the `train_gpu.py` module made it incompatible. The problem is described in issue https://github.com/facebookresearch/PyTorch-BigGraph/issues/162.

## How Has This Been Tested (if it applies)

I run all the tests on my machine and also trained a model with GPU on Google Colab. Colab only has Python 3.6 available and it did not throw any errors.

## Checklist

- [x] The documentation is up-to-date with the changes I made.
- [x] I have read the **CONTRIBUTING** document and completed the CLA (see **CONTRIBUTING**).
- [x] All tests passed, and additional code has been covered with new tests.

Pull Request resolved: https://github.com/facebookresearch/PyTorch-BigGraph/pull/163

Reviewed By: adamlerer

Differential Revision: D24502253

Pulled By: lw

fbshipit-source-id: ea7cceadb20a080bd77b095faf61ad73ed4a16b1