Fix Pytorch regressions in filament GPU (cudart and embeddings CPU grad)
Summary:
Pytorch changes have caused a regression in GPU Filament. This fixes them.

The new cudart code won't work with old versions of PyTorch due to the bug in the cudart signature, but it works for me internally and with 1.4.0. If I can isolate a version that this fails with, I could add a workaround path that reverts to the ctypes code.

Reviewed By: lw

Differential Revision: D24823912

fbshipit-source-id: a07b873cd8a93793eb59ea27c5b07c57dde318f3