fix the number of embedding_storage for unpartitioned entities that exist in both lhs and rhs
Summary: Allocating shared memory is expensive. We do this once at the beginning and then cache these storages and re-use them repeatedly during training. hence we need to estimate how many storages we need and the case of 1x1 partitions is handled poorly (because, it's rather rare) and the code decides it needs two storages for the those entities that appear on both sides but are alwasys the same partition on both sides. So only one storage is needed.

Reviewed By: lw

Differential Revision: D25179561

fbshipit-source-id: 1b3f9b896efb6127bcf03e8154bc26e8875e8553